requirements.txt
opencv-python
sqlalchemy
pillow
requests
Jinja2
tkinterweb
tqdm
pynput
cryptography
